requirements.txt
dash
plotly
pandas
numpy
scikit-learn==1.4.2
catboost
xgboost
gunicorn
dash-bootstrap-components
joblib
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
python-dotenv==1.0.0